// backup-src/types.ts
export interface LocationProps {
  id: string;
  distance: number;
  address: string;
}
// ...
export enum CategoryEnum {
  house = "house",
  shortlet="shortlet",
  hotel = "hotel",
  office = "office",
  land = "land",
  shop = "shop",
  others = "others"
}

export enum ListForEnum {
  rent = "rent",
  sale = "sale"
}